Revert "Moved aws s3 annotation."
This reverts commit 6cb17cb40bfbda4a5fd5da44db7511eb241a6df2.
